twoInports1Cube/requirements.txt
# OE Stack
openeye-toolkits
openeye-floe==0.2.148
